Add support in StaticImage
